refactor: Set "ydb" id in all providers
