+++ keycloak-ydb-extension/docker/conf/keycloak.conf
@@ -3,8 +3,8 @@
 
 # --- YDB / JPA (this extension) ---
 # JDBC URL for YDB (required). Override with env for different hosts, e.g.:
-spi-connections-jpa-default-ydb-jdbc-url=jdbc:ydb:grpc://ydb:2136/local
-spi-connections-jpa-default-show-sql=true
+spi-connections-jpa-ydb-jdbc-url=jdbc:ydb:grpc://ydb:2136/local
+spi-connections-jpa-ydb-show-sql=true
 
 # YDB Liquibase provider: index creation threshold (optional, default 300000)
 # spi-connections-liquibase-ydb-liquibase-index-creation-threshold=300000
