+++ keycloak-ydb-extension/src/main/kotlin/tech/ydb/keycloak/config/ProviderConfig.kt
@@ -1,5 +1,6 @@
 package tech.ydb.keycloak.config
 
-object ProviderPriority {
+object ProviderConfig {
   const val PROVIDER_PRIORITY = Int.MAX_VALUE
+  const val PROVIDER_ID = "ydb"
 }
