+++ keycloak-ydb-extension/src/main/kotlin/tech/ydb/keycloak/realm/YdbRealmProviderFactory.kt
@@ -6,7 +6,8 @@ import org.keycloak.models.KeycloakSession
 import org.keycloak.models.KeycloakSessionFactory
 import org.keycloak.models.RealmProviderFactory
 import org.keycloak.connections.jpa.JpaConnectionProvider
-import tech.ydb.keycloak.config.ProviderPriority.PROVIDER_PRIORITY
+import tech.ydb.keycloak.config.ProviderConfig.PROVIDER_ID
+import tech.ydb.keycloak.config.ProviderConfig.PROVIDER_PRIORITY
 
 class YdbRealmProviderFactory() : RealmProviderFactory<YdbRealmProvider> {
 
@@ -34,11 +35,7 @@ class YdbRealmProviderFactory() : RealmProviderFactory<YdbRealmProvider> {
     // no operations
   }
 
-  override fun getId(): String = ID
+  override fun getId(): String = PROVIDER_ID
 
   override fun order(): Int = PROVIDER_PRIORITY
-
-  private companion object {
-    private const val ID = "ydb-realm-provider-factory"
-  }
 }
